Added a new custom implementation of SHA256 so we can get the midstate. The block size was changed to 80 bytes, this changes were made so we could be compatinle with current bitcoin asic minner hardware.

+++ src/main/java/com/thunderbolt/mining/StandardMiner.java
@@ -122,7 +122,7 @@ public class StandardMiner implements IMiner
                 }
             }
 
-            block.getHeader().setTimeStamp(OffsetDateTime.now(ZoneOffset.UTC).toEpochSecond());
+            block.getHeader().setTimeStamp((int)OffsetDateTime.now(ZoneOffset.UTC).toEpochSecond());
             block.getHeader().setBits(m_blockchain.computeTargetDifficulty());
             block.getHeader().setParentBlockHash(m_blockchain.getChainHead().getHeader().getHash());
